moved exception namespace
